feat: add exception handling and ResourceNotFoundException
Having a global exception handler should simplify the exception handling. All exceptions that should be communicated to the client will have to implement the ApiLevelException interface.

The ApiError class serves as a way of controlling what information is passed to the client. The client is going to know the redundant data (such as status code or path) but stating it explicitly should make it easier in further development.