feat: add builder and constructor lombok annotations to entities for tests
During testing, it's convenient to have builder methods for creating required objects concisely. Tests are already quite long and can be hard to read, so I'm looking for any way to simplify them. This may seem quite heavy-handed because this generates a lot of code just for tests in the src directory. Builders may be useful in the future when creating resources, but if they will not be used, they should be moved to the test package.