feat: change the cinema model by adding a Row entity
Previous model (room -> seats) was limited to rooms with simple structure, because there was only one field specifying the row. According to the task specification, we have to be able to check whether there are 2 reserved places in a row. On previous model, this operation would be very costly, since we would have to filter seats with specified row. With the new model (room -> rows -> seats) it will be a non-problem.

Another benefit of this model is that in the future, the row entity can be extended to store information about distance from screen.